% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs990_lookup.R
\name{get_990_data}
\alias{get_990_data}
\title{Lookup IRS 990 Data Using EINs}
\usage{
get_990_data(x)
}
\arguments{
\item{x}{user-supplied csv file via the file.choose() function}
}
\value{
csv file with data from IRS 990 xml files corresponding with user-supplied EINs
}
\description{
Lookup IRS 990 Data Using EINs
}
\examples{
# simple run the following function call and select a csv file with one column of EINs
# please include a header row such as "eins"
get_990_data()

}
\author{
Michael Pawlus, \email{michael.pawlus@gmail.com}
}
